+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 0.0.3
+* Fix watch mode to work with both files and dirs (using watch-tree)
 
 ### 0.0.2
 * Refactor lib modules